Improve input/output files

+++ solutions/beecrowd/1099/1099.cpp
@@ -2,7 +2,7 @@
 #include <cstdio>
 
 int main() {
-    std::int16_t n, i, f, x, j, s;
+    std::int32_t n, i, f, x, j, s;
 
     scanf("%d", &n);
